function-app/requirements.txt
azure-functions
openai>=1.0.0
python-dotenv
requests
azure-identity
azure-storage-blob
azure-ai-documentintelligence